refactor: 제네레이터 생성기 위치 이동
: 도메인 레벨로 패키지 이동
: 해당 IdGenerator는 특정 도메인에 국한되지 않고, 재사용이 가능하므로 domain 레벨로 이동
